requirements.txt
# === core ===
Django == 4.1.5
djangorestframework == 3.14
djangorestframework-simplejwt == 5.2.2
django-filter == 22.1
django-money == 3.0.0
psycopg2-binary == 2.9.5
django-crontab == 0.7.1

# === API docs ===
drf-yasg == 1.21.4

# === tests ===
model-bakery == 1.9.0
parameterized == 0.8.1
coverage == 7.0.5

# === linting ===
black == 22.12.0
isort == 5.11.4
flake8 == 5.0.4
pyproject-flake8==5.0.4